fixup! getTablesConfigDefinition
